Revise repository interfaces
Breaking change. Corrected the interface methods to proper naming (e.g.
findEventsByUser'Since' -> findEventsByUser'Before'). This is because
the intent of the methods is to return entities that existed 'before'
a given time. 'Since' is a very confusing term to use in that context.

Also, removed unnecessary methods from the interface (e.g. findEventsByUser).
The reason behind this is that there will be no context to the returned
data. The data may be randomly ordered. The fact that the parameters
include a limit makes this "get-all" type of method obsolete. Similar
"get-all" behavior can be achieved effectively by using the other
"find-before" methods and supplying the current time as a parameter.