Add a isFinished flag to EventImpl in data-objectify
Added a flag that is being checked every time an event is retrieved. If the event's end
date is due but the isFinished flag is set to false, the event will be updated to its correct
flag value. This is to help lessen the amount of data being returned by geo-spatial
queries since it can only be done along with equality filters (i.e. filter events whose
flag is set to false).