Create welcome and sign-in pages
Created the HTML files for Welcome and Sign in. Changed A LOT of configuration
in Gruntfile.js to be able to package the entire app into dist/ such that
running a server on that directory will yield a shippable app. The side-effect
is that adding CSS files into the HTML page must be done via the Gruntfile.

Hopefully what I'm doing is scalable.