Implement Facebook authentication/registration
A user can now authentication via Facebook. If the user already has an account,
the facebook identity will be linked to this account when applicable. The application
will first verify the access token provided and that the token is meant for us, is not 
expired, and is owned by the same user that was provided in the authentication request.