Rename SessionService to AuthenticationTokenService
Since we will be using JWT, it would be strange to use sessions as well. I've researched about IP spoofing and am confident enough that a transaction done with a stolen token will have no benefit to the spoofer.

I'll leave the Session related classes in there just in case but for now, they're unused.