Add capability to add a marker in the map.
The user can now place markers by expanding the map, clicking the "place marker"
button, and then clicking anywher in the map. The next step might be to add a tooltip
that instructs the user to click on the map. As of writing, the only behavior is that the 
button becomes disabled.