Switch REST API Implementation to Swagger
Cloud Endpoint is promising but, it sucks everywhere so far. The UI is bad, authentication
is pretty coupled with Google OAuth. It's not very customizable. For now, Swagger
serves my needs with less roadblocks.

Authentication is now handled by a Servlet Filter that sets the UserPrincipal of the
HttpServletRequest. The Principal being set is an instance of UserPrincipal which has a getter
for a User domain class. To enforce this in an API operation, the annotation
@Authenticated should be used. This checks if the Principal in the HttpServletRequest
is an instance of UserPrincipal and if the getUser() returns a non-null value.